% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBatchParameters.R
\name{calculateBatch}
\alias{calculateBatch}
\title{Calculate clustering evaluation indicators for simulated data}
\usage{
calculateBatch(mat, batch_info, TSNE_ntop, PCA_ntop)
}
\arguments{
\item{mat}{Gene expression matrix}

\item{batch_info}{Cell Batch Label}

\item{TSNE_ntop}{A numerical scalar that specifies the number of features with the highest variance used for TSNE dimensionality reduction.}

\item{PCA_ntop}{A numerical scalar that specifies the number of features with the highest variance used for PCA dimensionality reduction.}
}
\value{
The value or list of clustering evaluation
}
\description{
Calculate clustering evaluation indicators for simulated data
}
