% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBatchParameters.R
\name{calculateBatchParameters}
\alias{calculateBatchParameters}
\title{Calculate batch effect evaluation indicators and normalize.}
\usage{
calculateBatchParameters(SCEList, TSNE_ntop = 500, PCA_ntop = 500)
}
\arguments{
\item{SCEList}{SCEList is a list containing multiple SingleCellExperiment objects.Please ensure that each SCE object has a 'Batch' column in its colData.}

\item{TSNE_ntop}{A numerical scalar that specifies the number of features with the highest variance used for TSNE dimensionality reduction.}

\item{PCA_ntop}{A numerical scalar that specifies the number of features with the highest variance used for PCA dimensionality reduction.}
}
\value{
Batch effect evaluation list and normalized values
}
\description{
This method calculates batch effects for each SCE object and normalizes the indicators. The higher the normalization value, the more prominent the batch effect.
}
\examples{
data("splatter_sce_Batch", "SPARSim_sce_Batch")
SCEList <- list(SPARSim = SPARSim_sce_Batch, splatter=splatter_sce_Batch)
test <- calculateBatchParameters(SCEList)


}
