% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateClusterParameters.R
\name{calculateCluster}
\alias{calculateCluster}
\title{Calculate clustering evaluation indicators}
\usage{
calculateCluster(mat, cluster_info, TSNE_ntop, PCA_ntop)
}
\arguments{
\item{mat}{Gene expression matrix}

\item{cluster_info}{Cell group label}

\item{TSNE_ntop}{A numerical scalar that specifies the number of features with the highest variance used for TSNE dimensionality reduction.}

\item{PCA_ntop}{A numerical scalar that specifies the number of features with the highest variance used for PCA dimensionality reduction.}
}
\value{
The value or list of clustering evaluation
}
\description{
Calculate clustering evaluation indicators
}
