% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateClusterParameters.R
\name{calculateClusterParameters}
\alias{calculateClusterParameters}
\title{Calculate the similarity in clustering evaluation metrics between the first dataset and other datasets}
\usage{
calculateClusterParameters(SCEList, TSNE_ntop = 500, PCA_ntop = 500)
}
\arguments{
\item{SCEList}{SCEList is a list containing multiple SingleCellExperiment objects.Please ensure that each SCE object has a 'group' column in its colData.}

\item{TSNE_ntop}{A numerical scalar that specifies the number of features with the highest variance used for TSNE dimensionality reduction.}

\item{PCA_ntop}{A numerical scalar that specifies the number of features with the highest variance used for PCA dimensionality reduction.}
}
\value{
The value and list of clustering evaluation
}
\description{
This method calculates the similarity between the clustering index of the first SCE object and the clustering indices of other SCE objects,
and normalizes the similarity to obtain the clustering similarity score between the first SCE object and other SCE objects.

The calculated score represents the similarity in clustering performance between the original dataset and the simulated dataset,
with higher scores indicating greater similarity between the original and simulated data.
}
\examples{
data("ESCO_sce_293T_jurkat", "muscat_sce_293T_jurkat", "scDesign3_sce_293T_jurkat", "origin_sce_293T_jurkat")
SCEList <- list(origin=origin_sce_293T_jurkat, ESCO=ESCO_sce_293T_jurkat, muscat=muscat_sce_293T_jurkat, scDesign3 = scDesign3_sce_293T_jurkat)
test <- calculateClusterParameters(SCEList)

}
