% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDEGs.R
\name{calculateDEGs}
\alias{calculateDEGs}
\title{Calculate the number of DEGs between groups}
\usage{
calculateDEGs(expr_data, group, P.Value, logFC, method)
}
\arguments{
\item{expr_data}{Gene expression matrix}

\item{group}{cell group}

\item{P.Value}{P.Value threshold for screening differentially expressed genes}

\item{logFC}{LogFC threshold for screening differentially expressed genes}
}
\value{
the number of DEGs
}
\description{
Calculate the number of DEGs between groups
}
