% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDEGs.R
\name{calculateDEGsParameters}
\alias{calculateDEGsParameters}
\title{Calculate the difference in DEGs between the first dataset and other datasets}
\usage{
calculateDEGsParameters(SCEList, P.Value = 0.05, logFC = 1, method = "GLM")
}
\arguments{
\item{SCEList}{SCEList is a list containing multiple SingleCellExperiment objects. Please ensure that each SCE object has a 'group' column in its colData, and that the number of types in each SCE object group is the same.}

\item{P.Value}{P.Value threshold for screening differentially expressed genes}

\item{logFC}{LogFC threshold for screening differentially expressed genes}

\item{method}{The method of using edgeR for DEA, method must be GLM or QLF.}
}
\value{
List of DEG quantities
}
\description{
This method calculates the differentially expressed genes of different groups between the first SCE object and each other SCE object.
The calculated score represents the difference in the number of differentially expressed genes between the original dataset and the simulated dataset.
The lower the score, the more similar the original and simulated data are
}
\examples{
data("ESCO_sce_293T_jurkat", "muscat_sce_293T_jurkat", "scDesign3_sce_293T_jurkat", "origin_sce_293T_jurkat")
SCEList <- list(origin=origin_sce_293T_jurkat, ESCO=ESCO_sce_293T_jurkat, muscat=muscat_sce_293T_jurkat, scDesign3 = scDesign3_sce_293T_jurkat)
test <- calculateDEGsParameters(SCEList)

}
