% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDataProperty.R
\name{StandardizedCalculationScore}
\alias{StandardizedCalculationScore}
\title{Standardize the output of the calcDataProperty function to calculate scores}
\usage{
StandardizedCalculationScore(tmp_list)
}
\arguments{
\item{tmp_list}{The result returned by the calcDataProperty function}
}
\value{
The standardized results
}
\description{
Standardize the output of the calcDataProperty function to calculate scores.
}
