% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBatchParameters.R
\name{calculateBatchParameters}
\alias{calculateBatchParameters}
\title{Calculate batch effects and return the results as a list}
\usage{
calculateBatchParameters(SCEList, TSNE_ntop = 500, PCA_ntop = 500)
}
\arguments{
\item{SCEList}{SCEList is a list containing multiple SingleCellExperiment objects.Please ensure that each SCE object has a 'Batch' column in its colData.}

\item{TSNE_ntop}{A numerical scalar that specifies the number of features with the highest variance used for TSNE dimensionality reduction.}

\item{PCA_ntop}{A numerical scalar that specifies the number of features with the highest variance used for PCA dimensionality reduction.}
}
\value{
List of Batch Effect evaluation
}
\description{
Calculate batch effects and return the results as a list
}
\examples{
data("splatter_sce_Batch", "SPARSim_sce_Batch")
SCEList <- list(SPARSim = SPARSim_sce_Batch, splatter=splatter_sce_Batch)
test <- calculateBatchParameters(SCEList)


}
